Move alert extension to a separate file
